backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
firebase-admin==6.3.0
httpx==0.26.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0
geopy==2.4.1
pillow==10.2.0
groq>=0.11.0
geohash2==1.1
